🏗️  Implement Main view extend, move Seat flow

+++ frontend/static/views/SelectView.js
@@ -1,6 +1,8 @@
 import AbstractView from "./AbstractView.js";
 import NavComponent from "../js/common/nav.js";
 import initSeats from "../js/seat-selection/seat-selection.js";
+import sessionStorageToObj from "../js/common/sessionStorageToObj.js";
+import toast from "../js/common/toast.js";
 
 const Price = {
   oneday: {
@@ -11,7 +13,7 @@ const Price = {
   },
   charge: {
     50: 50000,
-    100: 10000,
+    100: 100000,
   },
 };
 
@@ -105,53 +107,54 @@ export default class extends AbstractView {
     if (prevPath === "login") {
       prevBtn.setAttribute("href", "/ticket");
 
-      const ticketObj = sessionStorage
-        .getItem("ticket")
-        .split("{")[1]
-        .split("}")[0]
-        .split(",")
-        .reduce((obj, el) => {
-          const [key, val] = el.split(":").map((str) => str.replace(/\"/g, ""));
-          obj[key] = val;
-          return obj;
-        }, {});
+      const ticketObj = sessionStorageToObj("ticket");
+      const { time, auth } = ticketObj;
+      const formattedPrice = `${Intl.NumberFormat("ko-KR").format(
+        Price[auth][time]
+      )}원`;
 
-      const { time, auth, history } = ticketObj;
       ticketInfoArr[0].innerText = auth === "oneday" ? "당일권" : "시간권";
       ticketInfoArr[1].innerText = `${time}시간`;
-      totalPrice.innerText = `${Price[auth][time]}원`;
+      totalPrice.innerText = formattedPrice;
 
       payBtn.addEventListener("click", (e) => {
         const selectedSeat = sessionStorage.getItem("lastSelected");
+        //선택한 좌석이 있을 경우
         if (!selectedSeat) {
-          alert("좌석을 선택해주세요!");
-          //넘어가기 방지
+          payBtn.parentElement.setAttribute("href", "/select");
+          // toast("좌석을 선택해주세요!");
+          alert("좌석을 선택해주세요");
         } else {
           //validation - 좌석 data fetch받아와서 자리 여전히 없으면 req 넘기고, 다음 페이지로 이동
-          const req = {
-            category: auth,
-            duration: time,
-            price: Price[auth][time],
-            table: 1,
-            position: Number(selectedSeat),
-          };
-          //api.post(req)
-          console.log(req);
-          sessionStorage.clear();
-          sessionStorage.setItem("history", "using");
+          if (true) {
+            payBtn.parentElement.setAttribute("href", "/paycheck");
+            const req = {
+              category: auth,
+              duration: time,
+              price: Price[auth][time],
+              table: sessionStorage.getItem("table"),
+              position: Number(selectedSeat.replace(/[^0-9]/g, "")),
+              //현재 시각
+            };
+            //api.post(req)
+            console.log("selected seta", req);
 
-          //else
-          // 다음페이지로 넘어가기 방지
-          // alert 좌석을 다시 선택해주세요
-          // 자리 re-render
-          // localstorage에서 lastSelected 초기화
+            sessionStorage.clear();
+            sessionStorage.setItem("history", "using");
+            sessionStorage.setItem("price", formattedPrice);
+          } else {
+            alert("좌석을 다시 선택해주세요");
+            payBtn.parentElement.setAttribute("href", "/select");
+            localStorage.removeItem("table");
+            localStorage.removeItem("lastSelected");
+          }
         }
       });
     }
-    //(2) 퇴실 Main에서 온 경우
+    //퇴실 Main에서 버튼 선택
     else if (prevPath === "before") {
       prevBtn.setAttribute("href", "/main"); //퇴실메인으로 표시해야 함
-      //좌석 선택
+      //좌석 선택 - 퇴실뷰 나오면 그때 구현!!!
       if (path === "select") {
         sessionStorage.clear();
         sessionStorage.setItem("history", "main");
@@ -160,9 +163,10 @@ export default class extends AbstractView {
       else if (path === "extend") {
         const disableSelect = document.getElementById("disable--seat-view");
         const warningMessage = document.createElement("h1");
+        const ticketType = document.querySelector(".seat__selected");
 
         seatTitle.innerText = "Your Selected Ticket";
-        disableSelect.style.visibility = "visible";
+        ticketType.innerText = disableSelect.style.visibility = "visible";
 
         warningMessage.innerText =
           "좌석 선택은 이용권 연장 완료 후 시도해 주시기 바랍니다.";
@@ -171,10 +175,14 @@ export default class extends AbstractView {
         warningMessage.style.padding = "2em";
         disableSelect.append(warningMessage);
 
+        prevBtn.addEventListener("click", () => {});
+
         sessionStorage.clear();
         sessionStorage.setItem("history", "before");
       }
-    } else if (prevPath === "using") {
+    }
+    // 사용중 Main에서 버튼 선택
+    else if (prevPath === "using") {
       // &&시간이 남아있는 경우에만 자리이동 가능(이용중 메인에서 온 경우)
       if (path === "move") {
         // 왼편 결제하기 컴포넌트 지우기, 결제하기버튼 -> 이동하기 버튼으로 바꾸기
@@ -187,8 +195,40 @@ export default class extends AbstractView {
         $usingTime.innerText = "";
         paymethod.innerText = "시간권 차감";
 
-        sessionStorage.clear();
-        sessionStorage.setItem("history", "using");
+        payBtn.addEventListener("click", (e) => {
+          const selectedSeat = sessionStorage.getItem("lastSelected");
+          //선택한 좌석이 있을 경우
+          if (!selectedSeat) {
+            payBtn.parentElement.setAttribute("href", "/select");
+            // toast("좌석을 선택해주세요!");
+            alert("좌석을 선택해주세요");
+          } else {
+            //validation - 좌석 data fetch받아와서 자리 여전히 없으면 req 넘기고, 다음 페이지로 이동
+            if (true) {
+              payBtn.parentElement.setAttribute("href", "/paycheck");
+              const req = {
+                category: auth,
+                duration: time,
+                price: Price[auth][time],
+                table: sessionStorage.getItem("table"),
+                position: Number(selectedSeat.replace(/[^0-9]/g, "")),
+                //현재 시각
+              };
+              //api.post(req)
+              console.log("selected seta", req);
+
+              sessionStorage.clear();
+              sessionStorage.setItem("history", "using");
+              sessionStorage.setItem("price", formattedPrice);
+              sessionStorage.setItem("moved", selectedSeat);
+            } else {
+              alert("좌석을 다시 선택해주세요");
+              payBtn.parentElement.setAttribute("href", "/select");
+              localStorage.removeItem("table");
+              localStorage.removeItem("lastSelected");
+            }
+          }
+        });
       }
       //이용중메인 -> 시간만 연장
       else if (path == "extend") {
@@ -202,10 +242,12 @@ export default class extends AbstractView {
           "좌석 선택은 이용권 연장 완료 후 시도해 주시기 바랍니다.";
         warningMessage.style.color = "white";
         warningMessage.style.fontSize = "50px";
+        warningMessage.style.margin = "2em";
         disableSelect.append(warningMessage);
 
         sessionStorage.clear();
         sessionStorage.setItem("history", "using");
+        sessionStorage.setItem("price", formattedPrice);
       }
     }
   }
